with this commit 

- reuse stream of cupoch
- attach stream to mem using `cudaStreamAttachMemAsync`

the performance has been improved with the following result.

|   | h2d / h2h time(ms)| stable? |
| ------------- | ------------- |---------|
| pcl |  4ms | stable| 
| CudaDefaultAllocation  | 6-8ms |unstable  |
|PoolAllocation  | 13ms | stable |
|CudaManagedMemory  | 6.8ms | stable |
|CudaManagedMemoryPool  | 13ms | stable |

_Originally posted by @ZhenshengLee in https://github.com/ZhenshengLee/perception_cupoch/issues/9#issuecomment-901923622_

Closes #9